--- src/main/java/com/springboot/connectmate/constants/Constants.java
@@ -1,6 +1,0 @@
-package com.springboot.connectmate.constants;
-
-public class Constants {
-    public static final String WELCOME_MESSAGE = "%s joined to chat";
-    public static final String DISCONNECT_MESSAGE = "%s disconnected";
-}
